refactor dpm wave action routes
