+++ src/app/router_registry.py
@@ -35,6 +35,7 @@ from app.routers.dpm_command_center_portfolio_memory import (
 )
 from app.routers.dpm_construction import router as dpm_construction_router
 from app.routers.dpm_proof_packs import router as dpm_proof_packs_router
+from app.routers.dpm_wave_actions import router as dpm_wave_actions_router
 from app.routers.dpm_wave_ai import router as dpm_wave_ai_router
 from app.routers.dpm_wave_campaign_definitions import (
     router as dpm_wave_campaign_definitions_router,
@@ -115,6 +116,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_construction_router)
     app.include_router(dpm_proof_packs_router)
     app.include_router(dpm_waves_router)
+    app.include_router(dpm_wave_actions_router)
     app.include_router(dpm_wave_evidence_router)
     app.include_router(dpm_wave_ai_router)
     app.include_router(workbench_router)
